rename to LocalBlockOutStream
