+++ clients/unshaded/src/main/java/tachyon/client/RemoteBlockInStream.java
@@ -84,7 +84,7 @@ public class RemoteBlockInStream extends BlockInStream {
   /**
    * If we are re-caching the file, we write it to a block out stream as we read it.
    */
-  private BlockOutStream mBlockOutStream = null;
+  private LocalBlockOutStream mLocalBlockOutStream = null;
 
   /**
    * The under filesystem configuration that we use to set up the checkpoint input stream
@@ -145,8 +145,8 @@ public class RemoteBlockInStream extends BlockInStream {
   private void cancelRecache() throws IOException {
     if (mRecache) {
       mRecache = false;
-      if (mBlockOutStream != null) {
-        mBlockOutStream.cancel();
+      if (mLocalBlockOutStream != null) {
+        mLocalBlockOutStream.cancel();
       }
     }
   }
@@ -156,12 +156,12 @@ public class RemoteBlockInStream extends BlockInStream {
     if (mClosed) {
       return;
     }
-    if (mRecache && mBlockOutStream != null) {
+    if (mRecache && mLocalBlockOutStream != null) {
       // We only finish re-caching if we've gotten to the end of the file
       if (mBlockPos == mBlockInfo.length) {
-        mBlockOutStream.close();
+        mLocalBlockOutStream.close();
       } else {
-        mBlockOutStream.cancel();
+        mLocalBlockOutStream.cancel();
       }
     }
     if (mCheckpointInputStream != null) {
@@ -205,9 +205,10 @@ public class RemoteBlockInStream extends BlockInStream {
     int bytesLeft = len;
     // Lazy initialization of the out stream for caching to avoid collisions with other caching
     // attempts that are invalidated later due to seek/skips
-    if (bytesLeft > 0 && mBlockOutStream == null && mRecache) {
+    if (bytesLeft > 0 && mLocalBlockOutStream == null && mRecache) {
       try {
-        mBlockOutStream = new BlockOutStream(mFile, WriteType.TRY_CACHE, mBlockIndex, mTachyonConf);
+        mLocalBlockOutStream = new LocalBlockOutStream(mFile, WriteType.TRY_CACHE, mBlockIndex,
+            mTachyonConf);
       } catch (IOException ioe) {
         LOG.warn("Recache attempt failed.", ioe);
         cancelRecache();
@@ -220,7 +221,7 @@ public class RemoteBlockInStream extends BlockInStream {
       int bytesToRead = (int) Math.min(bytesLeft, mCurrentBuffer.remaining());
       mCurrentBuffer.get(b, off, bytesToRead);
       if (mRecache) {
-        mBlockOutStream.write(b, off, bytesToRead);
+        mLocalBlockOutStream.write(b, off, bytesToRead);
       }
       off += bytesToRead;
       bytesLeft -= bytesToRead;
@@ -247,7 +248,7 @@ public class RemoteBlockInStream extends BlockInStream {
           return len - bytesLeft;
         }
         if (mRecache) {
-          mBlockOutStream.write(b, off, readBytes);
+          mLocalBlockOutStream.write(b, off, readBytes);
         }
         off += readBytes;
         bytesLeft -= readBytes;
